Demo rework of SURF to speed up algorithm
Several small changes were made to prevent the same computations from
being made twice. Most of the property functions were changed to
internal class variables, for example. Another small speedup came from
precomputing all of the NaN entries and using a lookup table in the
evaluate_SURF function.

The biggest speed up comes from parallelizing the scoring function with
joblib. This implementation fixes the number of threads to -1, meaning
it will use as many threads as possible, but we will extend this
version to make that number a n_jobs parameter.

Currently, this code only parallelizes the scoring function on a
per-record basis. I suspect that we can expand this out to parallelize
it over the record *and* feature space, since the evaluate_SURF
function scores for once instance and one feature at a time.

Lastly, I also suspect that we can parallelize other sections of the
code to speed it up. For example, the distance matrix can likely be
parallelized, which we should take advantage of.