Parallelize distance function calculations
For the most part, we use scipy’s pdist function to very quickly
calculate the distance matrix. This function doesn’t require
parallelization.

However, in the case where the data matrix has missing values, we have
to write a custom distance function. In this case, we can parallelize
the function using joblib.